requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==4.1.2
django-appconf==1.0.5
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-imagekit==4.1.0
django-phonenumber-field==7.0.0
django-tables2==2.4.1
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.23.4
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.5.1
phonenumbers==8.12.57
pilkit==2.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
tabulate==0.9.0
timeago==1.0.16
tzdata==2022.5
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
